% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_region.R
\name{check_region}
\alias{check_region}
\title{Helper: Check that a region is valid}
\usage{
check_region(region)
}
\arguments{
\item{region}{A string input of a region name. Valid regions are "Africa", "Americas", "Eastern Mediterranean", "Europe", "South-East Asia", and "Western Pacific".}
}
\value{
A cleaned region name if the region is valid.
}
\description{
Helper: Check that a region is valid
}
