% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_year.R
\name{check_year}
\alias{check_year}
\title{Helper: Check that a year exists in the data}
\usage{
check_year(measles_data, year_chosen)
}
\arguments{
\item{measles_data}{A data frame of measles case data using load_data().}

\item{year_chosen}{A number value representing a year to check.}
}
\value{
The year if it exists in the data.
}
\description{
Helper: Check that a year exists in the data
}
