% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{clean_data}
\alias{clean_data}
\title{Helper: Change region names ad join country coordinates}
\usage{
clean_data(data)
}
\arguments{
\item{data}{measles dataset from load_data()}
}
\value{
A dataframe of measles data with regions renamed
}
\description{
Helper: Change region names ad join country coordinates
}
