% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_summary.R
\name{epi_linked_cases}
\alias{epi_linked_cases}
\title{Summarise measles cases for a region over time}
\usage{
epi_linked_cases(measles_data, region = "Africa")
}
\arguments{
\item{measles_data}{a dataframe of measles case data}

\item{region}{a string input of a region name}
}
\description{
Produce a summary output of the proportion of measles cases that are epidemiologically linked for a specified region over time
}
\examples{
epi_linked_cases(measles_data, "Africa")
}
