% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_summary.R
\name{epi_linked_cases_table}
\alias{epi_linked_cases_table}
\title{Summarise measles cases for a region over time}
\usage{
epi_linked_cases_table(measles_data, region = "Africa")
}
\arguments{
\item{measles_data}{A data frame of measles case data.}

\item{region}{A string input of a region name.}
}
\value{
A tibble with year and proportion of epidemiologically linked measles cases.
}
\description{
Produce a summary output of the proportion of measles cases that are
epidemiologically linked for a specified region over time.
}
\examples{
epi_linked_cases_table(load_data(), "Africa")
}
