% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cases.R
\name{plot_cases}
\alias{plot_cases}
\title{Plot measles and rubella cases over time}
\usage{
plot_cases(measles_data, start = 2012, end = 2024)
}
\arguments{
\item{measles_data}{a dataframe of measles and rubella cases}

\item{start}{start year to plot cases}

\item{end}{end year to plot cases}
}
\value{
ggplot plot of cases between given years
}
\description{
Plot measles and rubella cases over time
}
\examples{
plot_cases(load_data(), 2012, 2024)
}
