% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_gdp_cases_map}
\alias{plot_gdp_cases_map}
\title{Plot Countries}
\usage{
plot_gdp_cases_map(measles_data, year_chosen = 2024, top_n = 20)
}
\arguments{
\item{measles_data}{a dataframe of measles case data}

\item{year_chosen}{a year from 2012 to 2024}

\item{top_n}{number of countries to be displayed}
}
\value{
a Leaflet map
}
\description{
Produce a map of countries with lowest measles case rate and highest GDP in a given year
}
\examples{
plot_gdp_cases_map(measles_data, 2024, 20)
}
